Move DrawPanel to draw folder
