+++ data/__init__.py
@@ -1,0 +1,3 @@
+from .resolver import DataResolver
+
+__all__ = [ 'DataResolver' ]
